requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via langfuse
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   streamlit
    #   uvicorn
distro==1.9.0
    # via openai
durationpy==0.10
    # via kubernetes
fastapi==0.115.14
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-auth==2.40.3
    # via kubernetes
googleapis-common-protos==1.70.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.73.1
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langfuse
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jiter==0.10.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
kubernetes==33.1.0
    # via -r requirements.in
langchain==0.3.26
    # via -r requirements.in
langchain-core==0.3.66
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==0.3.26
    # via -r requirements.in
langchain-text-splitters==0.3.8
    # via langchain
langfuse==3.0.6
    # via -r requirements.in
langgraph==0.5.0
    # via -r requirements.in
langgraph-checkpoint==2.1.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.5.0
    # via langgraph
langgraph-sdk==0.1.72
    # via langgraph
langsmith==0.4.3
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
narwhals==1.44.0
    # via altair
numpy==2.3.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
openai==1.92.2
    # via
    #   -r requirements.in
    #   langchain-openai
opentelemetry-api==1.34.1
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.34.1
    # via langfuse
opentelemetry-exporter-otlp-proto-common==1.34.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.34.1
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.34.1
    # via opentelemetry-exporter-otlp
opentelemetry-proto==1.34.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.34.1
    # via
    #   langfuse
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.55b1
    # via opentelemetry-sdk
orjson==3.10.18
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==24.2
    # via
    #   altair
    #   langchain-core
    #   langfuse
    #   langsmith
    #   pytest
    #   streamlit
pandas==2.3.0
    # via
    #   -r requirements.in
    #   streamlit
pillow==11.2.1
    # via streamlit
pluggy==1.6.0
    # via pytest
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
    #   streamlit
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langfuse
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   kubernetes
    #   langchain
    #   langchain-core
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.4
    # via
    #   -r requirements.in
    #   kubernetes
    #   langchain
    #   langfuse
    #   langsmith
    #   opentelemetry-exporter-otlp-proto-http
    #   requests-oauthlib
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   kubernetes
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via langchain
starlette==0.46.2
    # via fastapi
streamlit==1.46.1
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   langchain-core
    #   streamlit
tiktoken==0.9.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via streamlit
tqdm==4.67.1
    # via openai
typing-extensions==4.14.0
    # via
    #   altair
    #   fastapi
    #   langchain-core
    #   openai
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   kubernetes
    #   requests
uvicorn==0.34.3
    # via -r requirements.in
websocket-client==1.8.0
    # via kubernetes
wrapt==1.17.2
    # via langfuse
xxhash==3.5.0
    # via langgraph
zipp==3.23.0
    # via importlib-metadata
zstandard==0.23.0
    # via langsmith
